pyproject.toml
[project]
name = "taixuan-translator"
version = "1.0.0"
description = "太玄智译 - Windows本地部署的Python桌面PDF精准翻译解决方案"
authors = [{name = "太玄计算机软件开发工作室"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
keywords = ["pdf", "translation", "ocr", "desktop-app", "pyqt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: Other/Proprietary License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    # ─── Web框架（API服务）────────────────────────────────────────────
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",

    # ─── UI Framework ─────────────────────────────────────────────────
    "PyQt6>=6.4.0",
    "PyQt6-Qt6>=6.4.0",

    # ─── PDF Processing ──────────────────────────────────────────────
    "PyMuPDF>=1.23.0",
    "pdfminer.six>=20221105",

    # ─── Document Generation ──────────────────────────────────────────
    "python-docx>=1.1.0",

    # ─── OCR & Formula Recognition ────────────────────────────────────
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.5.0",

    # ─── Table Extraction ────────────────────────────────────────────
    "camelot-py[cv]>=0.11.0",
    "tabula-py>=2.9.0",

    # ─── Translation APIs ────────────────────────────────────────────
    "openai>=1.0.0",
    "deepl>=1.16.0",

    # ─── Local LLM ──────────────────────────────────────────────────
    "ollama>=0.1.0",

    # ─── Data & Database ────────────────────────────────────────────
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # ─── ePub & Subtitles ────────────────────────────────────────────
    "ebooklib>=0.17.0",
    "lxml>=4.9.0",
    "pysubs2>=1.6.0",
    "beautifulsoup4>=4.12.0",

    # ─── Utilities ──────────────────────────────────────────────────
    "loguru>=0.7.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",

    # ─── Configuration ──────────────────────────────────────────────
    "pyyaml>=6.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "tomli-w>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-qt>=4.2.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.5.0",
]
build = [
    "pyinstaller>=6.0.0",
    "nuitka>=1.8.0",
]

[project.scripts]
taixuan-translator = "taixuan_translator.main:main"

[project.gui-scripts]
taixuan-translator-gui = "taixuan_translator.main:main"

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["taixuan_translator"]

[tool.hatch.build.targets.sdist]
include = ["/taixuan_translator", "/tests", "/docs"]

# Black formatting
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
)/
'''

# Ruff linting
[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.isort]
known-first-party = ["taixuan_translator"]

# MyPy type checking
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "fitz.*",
    "pdfminer.*",
    "paddleocr.*",
    "camelot.*",
    "tabula.*",
    "ebooklib.*",
    "pysubs2.*",
]
ignore_missing_imports = true

# Pytest
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "gui: marks tests that require GUI",
]

# Coverage
[tool.coverage.run]
source = ["taixuan_translator"]
branch = true
omit = ["tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]